+++ apps/docs/src/app/platform/component-docs/platform-table/platform-table-examples/platform-table-single-row-selection-example.component.ts
@@ -6,7 +6,7 @@ import {
     TableDataSource,
     TableRowSelectionChangeEvent,
     TableDataProvider,
-    TableState
+    TableState,
 } from '@fundamental-ngx/platform/table';
 
 @Component({
@@ -20,12 +20,21 @@ export class PlatformTableSingleRowSelectionExampleComponent {
         this.source = new TableDataSource(new TableDataProviderExample());
     }
 
+    trackBy(index: number, item: ExampleItem): number {
+        return item.id;
+    }
+
+    rowComparator(left: ExampleItem, right: ExampleItem): boolean {
+        return left.id === right.id;
+    }
+
     onRowSelectionChange(event: TableRowSelectionChangeEvent<ExampleItem>): void {
         console.log(event);
     }
 }
 
 export interface ExampleItem {
+    id: number;
     name: string;
     description: string;
     price: {
@@ -85,6 +94,7 @@ function getNestedValue<T extends {}>(key: string, object: T): any {
 // Example items
 const ITEMS: ExampleItem[] = [
     {
+        id: 1,
         name: '10 Portable DVD player',
         description: 'diam neque vestibulum eget vulputate',
         price: {
@@ -97,6 +107,7 @@ const ITEMS: ExampleItem[] = [
         verified: true
     },
     {
+        id: 2,
         name: 'Astro Laptop 1516',
         description: 'pede malesuada',
         price: {
@@ -109,6 +120,7 @@ const ITEMS: ExampleItem[] = [
         verified: true
     },
     {
+        id: 3,
         name: 'Astro Phone 6',
         description: 'penatibus et magnis',
         price: {
@@ -121,6 +133,7 @@ const ITEMS: ExampleItem[] = [
         verified: true
     },
     {
+        id: 4,
         name: 'Beam Breaker B-1',
         description: 'fermentum donec ut',
         price: {
@@ -133,6 +146,7 @@ const ITEMS: ExampleItem[] = [
         verified: false
     },
     {
+        id: 5,
         name: 'Beam Breaker B-2',
         description: 'sapien in sapien iaculis congue',
         price: {
@@ -144,6 +158,7 @@ const ITEMS: ExampleItem[] = [
         verified: true
     },
     {
+        id: 6,
         name: 'Benda Laptop 1408',
         description: 'suspendisse potenti cras in',
         price: {
@@ -156,6 +171,7 @@ const ITEMS: ExampleItem[] = [
         verified: true
     },
     {
+        id: 7,
         name: 'Bending Screen 21HD',
         description: 'nunc nisl duis bibendum',
         price: {
@@ -168,6 +184,7 @@ const ITEMS: ExampleItem[] = [
         verified: false
     },
     {
+        id: 8,
         name: 'Blaster Extreme',
         description: 'quisque ut',
         price: {
@@ -180,6 +197,7 @@ const ITEMS: ExampleItem[] = [
         verified: true
     },
     {
+        id: 9,
         name: 'Broad Screen 22HD',
         description: 'ultrices posuere',
         price: {
@@ -192,6 +210,7 @@ const ITEMS: ExampleItem[] = [
         verified: true
     },
     {
+        id: 10,
         name: 'Camcorder View',
         description: 'integer ac leo pellentesque',
         price: {
@@ -204,6 +223,7 @@ const ITEMS: ExampleItem[] = [
         verified: true
     },
     {
+        id: 11,
         name: 'Cepat Tablet 10.5',
         description: 'rutrum rutrum neque aenean auctor',
         price: {
@@ -215,6 +235,7 @@ const ITEMS: ExampleItem[] = [
         verified: true
     },
     {
+        id: 12,
         name: 'Ergo Mousepad',
         description: 'tortor duis mattis egestas',
         price: {
@@ -227,6 +248,7 @@ const ITEMS: ExampleItem[] = [
         verified: true
     },
     {
+        id: 13,
         name: 'Ergo Screen E-I',
         description: 'massa quis augue luctus tincidunt',
         price: {
@@ -239,6 +261,7 @@ const ITEMS: ExampleItem[] = [
         verified: true
     },
     {
+        id: 14,
         name: 'Ergo Screen E-II',
         description: 'orci eget',
         price: {
@@ -250,6 +273,7 @@ const ITEMS: ExampleItem[] = [
         verified: false
     },
     {
+        id: 15,
         name: 'Gaming Monster',
         description: 'cubilia curae',
         price: {
@@ -261,6 +285,7 @@ const ITEMS: ExampleItem[] = [
         verified: false
     },
     {
+        id: 16,
         name: 'Gaming Monster Pro',
         description: 'pharetra magna vestibulum aliquet',
         price: {