+++ e2e/wdio/core/pages/dialog.po.ts
@@ -6,7 +6,7 @@ export class DialogPo extends CoreBaseComponentPo {
     root = '#page-content';
 
     // example selectors
-    templateDialog = 'fd-dialog-open-template-example ';
+    templateDialog = 'fd-template-based-dialog-example ';
     componentDialog = 'fd-component-based-dialog-example ';
     objectDialog = 'fd-dialog-object-example ';
     stateDialog = 'fd-dialog-state-example ';
@@ -24,7 +24,9 @@ export class DialogPo extends CoreBaseComponentPo {
     button = 'button';
     dialogOutput = 'p';
     busyIndicator = 'fd-busy-indicator';
+    // works by the same way, just need use different according to specific dialogs
     dialogContainer = this.dialog + '.cdk-drag';
+    dialogContainer2 = this.dialog + '.cdk-drag-disabled';
     resizeHandle = this.dialog + '.fd-dialog__resize-handle';
     dialogItems = this.dialog + 'li';
     searchBar = this.dialog + 'input';