+++ libs/platform/src/lib/form/multi-combobox/pipes/highlight.pipe.ts
@@ -1,6 +1,6 @@
 import { Pipe, PipeTransform } from '@angular/core';
 
-import { MatchingStrategy } from '@fundamental-ngx/platform/shared';
+import { getMatchingStrategyStartsWithPerTermReqexp, MatchingStrategy } from '@fundamental-ngx/platform/shared';
 import { MultiComboboxConfig } from '../multi-combobox.config';
 
 // TODO: refactor it with ComboboxHighlightPipe from Combobox and SearchHighlightPipe (https://github.com/SAP/fundamental-ngx/issues/5333)
@@ -16,6 +16,10 @@ export class MultiComboboxHighlightPipe implements PipeTransform {
             return value;
         }
 
+        if (matchingStrategy === MatchingStrategy.STARTS_WITH_PER_TERM) {
+            return this._searchByStrategyStartsWithPerTerm(value, searchText);
+        }
+
         if (matchingStrategy === MatchingStrategy.STARTS_WITH) {
             return this._searchByStrategyStartsWith(value, searchText);
         }
@@ -25,6 +29,12 @@ export class MultiComboboxHighlightPipe implements PipeTransform {
         }
 
         return value;
+    }
+
+    private _searchByStrategyStartsWithPerTerm(value: string, searchText: string): string {
+        const reqexp = getMatchingStrategyStartsWithPerTermReqexp(searchText);
+
+        return value.replace(reqexp, `$1<strong>$2</strong>`);
     }
 
     private _searchByStrategyStartsWith(value: string, searchText: string): string {