+++ libs/platform/src/lib/form/multi-input/public_api.ts
@@ -3,3 +3,4 @@ export { PlatformMultiInputComponent } from './multi-input.component';
 export { PlatformMultiInputModule } from './multi-input.module';
 export { PlatformMultiInputMobileComponent } from './multi-input-mobile/multi-input-mobile.component';
 export { PlatformMultiInputMobileModule } from './multi-input-mobile/multi-input-mobile.module';
+export { MultiInputConfig } from './multi-input.config';